database/requirement.txt
torch
kafka-python